Move ImageConfig into GenConfig file
