--- firebaseai/src/ImageConfig.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6056c2a52efe845e28c64a2bb3ee3ff8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
